[#11024] DocDB: Enable -Wsign-compare and -Wdeprecated
Summary: This diff enables sign-compare and deprecated warnings, and fixes found issue in codebase.

Test Plan: Jenkins

Reviewers: mpolitov

Reviewed By: mpolitov

Subscribers: ybase, bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D14713

+++ src/yb/client/backup-txn-test.cc
@@ -418,7 +418,7 @@ TEST_F(BackupTxnTest, FlushSysCatalogAndDelete) {
   ASSERT_NO_FATALS(WriteData());
   auto snapshot_id = ASSERT_RESULT(snapshot_util_->CreateSnapshot(table_));
 
-  for (int i = 0; i != cluster_->num_masters(); ++i) {
+  for (size_t i = 0; i != cluster_->num_masters(); ++i) {
     auto tablet_peer = cluster_->mini_master(i)->tablet_peer();
     ASSERT_OK(tablet_peer->tablet()->Flush(tablet::FlushMode::kSync));
   }
